+++ tests/types/main.rs
@@ -2,8 +2,4 @@
 mod common;
 
 mod opaque;
-mod opaque_list;
-mod opaque_set;
 mod type_name;
-mod type_name_list;
-mod type_name_set;