+++ tests/types/type_name/list.rs
@@ -1,3 +1,5 @@
+//! Integration tests for TypeNameList
+
 use crate::common::*;
 use display_as_debug::types::TypeNameList;
 
@@ -10,7 +12,6 @@ mod short {
     test_fmt!(new, TypeNameList::<String, Short>::new(100), "{:?}", EXPECTED);
     test_fmt!(of, TypeNameList::<String, Short>::of(0..100), "{:?}", EXPECTED);
     test_fmt!(from, TypeNameList::<String, Short>::from(0..100), "{:?}", EXPECTED);
-    test_fmt!(display, TypeNameList::<String, Short>::of(0..100), "{}", EXPECTED);
     test_get!(len, TypeNameList::<String, Short>::new(100), TypeNameList::len, 100);
 }
 
@@ -23,6 +24,5 @@ mod full {
     test_fmt!(new, TypeNameList::<String, Full>::new(100), "{:?}", EXPECTED);
     test_fmt!(of, TypeNameList::<String, Full>::of(0..100), "{:?}", EXPECTED);
     test_fmt!(from, TypeNameList::<String, Full>::from(0..100), "{:?}", EXPECTED);
-    test_fmt!(display, TypeNameList::<String, Full>::of(0..100), "{}", EXPECTED);
     test_get!(len, TypeNameList::<String, Full>::new(100), TypeNameList::len, 100);
 }