+++ tests/types/type_name/set.rs
@@ -1,3 +1,5 @@
+//! Integration tests for TypeNameSet
+
 use crate::common::*;
 use display_as_debug::types::TypeNameSet;
 
@@ -10,7 +12,6 @@ mod short {
     test_fmt!(new, TypeNameSet::<String, Short>::new(100), "{:?}", EXPECTED);
     test_fmt!(of, TypeNameSet::<String, Short>::of(0..100), "{:?}", EXPECTED);
     test_fmt!(from, TypeNameSet::<String, Short>::from(0..100), "{:?}", EXPECTED);
-    test_fmt!(display, TypeNameSet::<String, Short>::of(0..100), "{}", EXPECTED);
     test_get!(len, TypeNameSet::<String, Short>::new(100), TypeNameSet::len, 100);
 }
 
@@ -23,6 +24,5 @@ mod full {
     test_fmt!(new, TypeNameSet::<String, Full>::new(100), "{:?}", EXPECTED);
     test_fmt!(of, TypeNameSet::<String, Full>::of(0..100), "{:?}", EXPECTED);
     test_fmt!(from, TypeNameSet::<String, Full>::from(0..100), "{:?}", EXPECTED);
-    test_fmt!(display, TypeNameSet::<String, Full>::of(0..100), "{}", EXPECTED);
     test_get!(len, TypeNameSet::<String, Full>::new(100), TypeNameSet::len, 100);
 }